Trip_CrewAI/requirements.txt
crewai
langchain
openai
python-dotenv
requests
unstructured
tiktoken
langchain-community
langchain-core